[FEATURE] 문제 종류에 따라서 푼 문제 저장하는 로직 구체화

See merge request s11-webmobile1-sub2/S11P12E202!185

+++ backend/user/src/main/java/com/scf/user/profile/domain/dto/kafka/SolvedProblemKafkaRequestDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class SolvedProblemKafkaRequestDto {
 
-    private int problemId; // 문제 아이디
+    private Long problemId; // 문제 아이디
     private boolean isCorrect; // 정답여부
     private String choice; // 주관식
     private Map<Integer, Integer> choiceText; // 빈칸, 객관식
